[EB-64] Add tests for sellers

+++ src/sellers/service.py
@@ -3,13 +3,13 @@ import os
 from app import exceptions
 from app.conf.upload import MEDIA_ROOT
 from sellers.models import Contact, Seller
+from upload.models import UploadedFile
 from upload.service import UploadService
 from users.models import User
 
 
 class SellerService:
     """ Seller service """
-    CONTACTS_NOT_SUPPLIED = 'Contacts are not supplied'
 
     def create_or_update_object(self, user: User, data: dict) -> None:
         """ Creates or updates object.
@@ -25,11 +25,6 @@ class SellerService:
         logo_url = data.pop('logo_url', None)
         upload_service = UploadService()
 
-        # todo: by some reason serializer couldn't make contacts field to be required;
-        #  here is workaround for now
-        if not contacts:
-            raise exceptions.NotFound(msg=self.CONTACTS_NOT_SUPPLIED)
-
         seller, _ = Seller.objects.update_or_create(
             seller=user,
             defaults={'name': data['name'],
@@ -42,5 +37,6 @@ class SellerService:
             seller.logo_url = upload_to
             seller.save()
 
-        Contact.objects.delete_all_by_seller(seller)
-        Contact.objects.bulk_create([Contact(seller=seller, **contact) for contact in contacts])
+        if contacts:
+            Contact.objects.delete_all_by_seller(seller)
+            Contact.objects.bulk_create([Contact(seller=seller, **contact) for contact in contacts])
